chore(web): add actions for login and registers
